requirements.txt
PyMuPDF
pymongo
langchain
sentence-transformers
qdrant-client
fastapi
uvicorn
python-dotenv
pytesseract
pdf2image
opencv-python
Pillow
pdfminer.six
regex
weaviate-client
langchain
numpy
ollama
faiss-cpu
scikit-learn
transformers
torch
